pyproject.toml
[tool.poetry]
name = "vector-search-backend"
version = "0.1.0"
description = ""
authors = ["tyndev <17089219+tyndev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
langchain = "^0.2.5"
qdrant-client = "^1.9.1"
langchain-community = "^0.2.5"
openai = "^1.35.1"
langchain-qdrant = "^0.1.1"
langchain-openai = "^0.1.8"
pdfplumber = "^0.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
